+++ src/_test/core/revert2/offer/revert__offer__0xA2.sol
@@ -6,9 +6,9 @@ import "../../../NuggftV1.test.sol";
 
 abstract contract revert__offer__0xA2 is NuggftV1Test {
     function test__revert__offer__0xA2__fail__desc() public {
-        (uint160 tokenId, , uint16 itemId, uint96 floor) = scenario_dee_has_sold_an_item();
+        (uint24 tokenId, , uint16 itemId, uint96 floor) = scenario_dee_has_sold_an_item();
 
-        uint160 charliesTokenId = scenario_charlie_has_a_token();
+        uint24 charliesTokenId = scenario_charlie_has_a_token();
 
         uint96 value = floor + 1 ether;
 
@@ -16,9 +16,9 @@ abstract contract revert__offer__0xA2 is NuggftV1Test {
     }
 
     function test__revert__offer__0xA2__pass__desc() public {
-        (uint160 tokenId, , uint16 itemId, uint96 floor) = scenario_dee_has_sold_an_item();
+        (uint24 tokenId, , uint16 itemId, uint96 floor) = scenario_dee_has_sold_an_item();
 
-        uint160 charliesTokenId = scenario_charlie_has_a_token();
+        uint24 charliesTokenId = scenario_charlie_has_a_token();
 
         uint96 value = floor + 1 ether;
